match parser improvements
