+++ app/src/main/java/com/ricardosfp/zapping/data/repository/implementation/ZappingRepositoryImpl.kt
@@ -8,6 +8,7 @@ import com.ricardosfp.zapping.data.repository.model.*
 import dagger.hilt.android.qualifiers.*
 import javax.inject.*
 
+// todo test
 @Singleton
 class ZappingRepositoryImpl @Inject constructor(
     @ApplicationContext context: Context
@@ -15,7 +16,8 @@ class ZappingRepositoryImpl @Inject constructor(
 
     // since this field is already part of a singleton and is only used in here, there is no
     // strong need to inject it
-    private val parser: Parser = Parser.Builder() //        .charset(StandardCharsets.ISO_8859_1)
+    private val parser: Parser = Parser.Builder()
+        //.charset(StandardCharsets.ISO_8859_1)
         .context(context).cacheExpirationMillis((1000 * 60 * 60 * 24).toLong()).build()
 
     override suspend fun getMatches(): GetMatchesRepositoryResponse {
@@ -24,7 +26,19 @@ class ZappingRepositoryImpl @Inject constructor(
             // the library automatically changes the context of the coroutine
             val channel = parser.getChannel(BuildConfig.ZAPPING_URL)
 
-            GetMatchesRepositoryResponseSuccess(channel)
+            // is it worth it putting this into a separate class for testing purposes?
+            val articleList = channel.articles.mapNotNull {
+                val date = it.pubDate
+                val title = it.title
+
+                if (date != null && title != null) {
+                    MyArticle(date, title)
+                } else {
+                    null
+                }
+            }
+
+            GetMatchesRepositoryResponseSuccess(articleList)
         }
         catch (th: Throwable) {
             GetMatchesRepositoryResponseError(th)
