+++ app/src/main/java/com/ricardosfp/zapping/data/repository/model/GetMatchesRepositoryResponse.kt
@@ -1,9 +1,8 @@
 package com.ricardosfp.zapping.data.repository.model
 
-import com.prof.rssparser.*
-
 sealed class GetMatchesRepositoryResponse
 
-class GetMatchesRepositoryResponseSuccess(val channel: Channel): GetMatchesRepositoryResponse()
+class GetMatchesRepositoryResponseSuccess(val articles: List<MyArticle>):
+    GetMatchesRepositoryResponse()
 
 class GetMatchesRepositoryResponseError(val throwable: Throwable): GetMatchesRepositoryResponse()
