+++ app/src/main/java/com/ricardosfp/zapping/data/repository/model/MyArticle.kt
@@ -1,0 +1,3 @@
+package com.ricardosfp.zapping.data.repository.model
+
+class MyArticle(val date: String, val title: String)
