--- app/src/main/java/com/ricardosfp/zapping/domain/article/ArticleParser.kt
@@ -1,9 +1,0 @@
-package com.ricardosfp.zapping.domain.article
-
-import com.prof.rssparser.*
-import com.ricardosfp.zapping.domain.model.*
-
-interface ArticleParser {
-
-    fun parse(article: Article) : Match?
-}
