+++ app/src/main/java/com/ricardosfp/zapping/domain/match/MatchParser.kt
@@ -1,0 +1,9 @@
+package com.ricardosfp.zapping.domain.match
+
+import com.ricardosfp.zapping.data.repository.model.*
+import com.ricardosfp.zapping.domain.model.*
+
+interface MatchParser {
+
+    fun parse(article: MyArticle): Match?
+}
