+++ app/src/main/java/com/ricardosfp/zapping/domain/match/MatchParserImpl.kt
@@ -1,0 +1,58 @@
+package com.ricardosfp.zapping.domain.match
+
+import com.ricardosfp.zapping.data.repository.model.*
+import com.ricardosfp.zapping.domain.model.*
+import com.ricardosfp.zapping.infrastructure.util.date.*
+import java.text.*
+import java.util.*
+import javax.inject.*
+
+// todo test this class
+@Singleton
+class MatchParserImpl @Inject constructor(
+    private val dateUtils: DateUtils,
+): MatchParser {
+
+    companion object {
+        private val dateFormat = SimpleDateFormat("E, d MMM yyyy HH:mm:ss", Locale.ENGLISH)
+    }
+
+    override fun parse(article: MyArticle): Match? {
+        try {
+            val date = dateUtils.parse(dateFormat, article.date)
+            val originalText = article.title
+
+            if (date != null) {
+                val parts = originalText.split(" - ")
+                if (parts.size == 3) {
+                    val teams = parts[0].split(" x ")
+                    if (teams.size == 2) {
+                        val homeTeam = teams[0]
+                        val awayTeam = teams[1]
+                        val channel = parts[2]
+
+                        return Match(homeTeam, awayTeam, date, channel, originalText)
+
+                    } else {
+                        // not two teams
+                        // todo report this error
+                    }
+                } else {
+                    // not the right number of string slices to extract
+                    // information from
+                    // todo report this error
+                }
+            } else {
+                // date null
+                // todo report this error
+            }
+        }
+        catch (ex: Exception) {
+            ex.printStackTrace()
+            // todo report this error
+        }
+
+        return null
+    }
+
+}
